exercise 1

+++ code/Record.cpp
@@ -2,7 +2,7 @@
 #include "Record.h"
 
 Record::Record(std::string GAME_DATE_EST,
-    int TEAM_ID_home,
+    long long TEAM_ID_home,
     int PTS_home,
     double FG_PCT_home, double FT_PCT_home, double FG3_PCT_home,
     int AST_home, int REB_home,
